src/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
Jinja2==2.10.1
PyYAML==5.3.1
requests==2.21.0
requests-oauthlib==1.2.0